revert(local-scheduling): drop SchedulableTaskManager
LocalSchedulerManager already auto-registers every tagged
SchedulableInterface when its start() runs (via @injectAll on the
Schedulable tag), so the extracted SchedulableTaskManager only added
indirection — and, because it injected the scheduler while the module's
onShutdown resolves the runtime server that injects it back, a needless
DI cycle risk.

The reusable entry point is start() itself: any command that wants to
drive the scheduler injects "LocalSchedulerInterface" and calls start();
LocalSchedulerRuntimeServer does exactly that for `pristine start`. No
separate registrar class is needed.

This restores the auto-registering manager, the simple two-arg runtime
server, and moves the tagged-task tests back into the manager spec.
