+++ packages/local-scheduling/src/managers/managers.ts
@@ -1,2 +1,1 @@
 export * from "./local-scheduler.manager";
-export * from "./schedulable-task.manager";
