+++ packages/local-scheduling/src/servers/local-scheduler.runtime-server.spec.ts
@@ -1,7 +1,6 @@
 import "reflect-metadata";
 import {LocalSchedulerRuntimeServer} from "./local-scheduler.runtime-server";
 import {LocalSchedulerInterface} from "../interfaces/local-scheduler.interface";
-import {SchedulableTaskManager} from "../managers/schedulable-task.manager";
 import {LogHandlerInterface} from "@pristine-ts/logging";
 
 const createLogHandlerMock = (): jest.Mocked<LogHandlerInterface> => ({
@@ -21,30 +20,22 @@ const createSchedulerMock = (): jest.Mocked<LocalSchedulerInterface> => ({
   stop: jest.fn().mockResolvedValue(undefined),
 });
 
-const createRegistrarMock = () => ({register: jest.fn()}) as unknown as jest.Mocked<SchedulableTaskManager>;
-
 describe("LocalSchedulerRuntimeServer", () => {
   it("exposes a stable name for pristine start diagnostics", () => {
-    const server = new LocalSchedulerRuntimeServer(createSchedulerMock(), createRegistrarMock(), createLogHandlerMock());
+    const server = new LocalSchedulerRuntimeServer(createSchedulerMock(), createLogHandlerMock());
     expect(server.name).toBe("local-scheduler");
   });
 
-  it("start() registers tagged tasks, then starts the scheduler", async () => {
+  it("start() starts the scheduler (arming every tagged task)", async () => {
     const scheduler = createSchedulerMock();
-    const registrar = createRegistrarMock();
-    const server = new LocalSchedulerRuntimeServer(scheduler, registrar, createLogHandlerMock());
-
+    const server = new LocalSchedulerRuntimeServer(scheduler, createLogHandlerMock());
     await server.start();
-
-    expect(registrar.register).toHaveBeenCalledTimes(1);
     expect(scheduler.start).toHaveBeenCalledTimes(1);
-    // Registration must happen before arming, so start() arms the tasks it just registered.
-    expect(registrar.register.mock.invocationCallOrder[0]).toBeLessThan(scheduler.start.mock.invocationCallOrder[0]);
   });
 
   it("stop() awaits the scheduler's graceful stop", async () => {
     const scheduler = createSchedulerMock();
-    const server = new LocalSchedulerRuntimeServer(scheduler, createRegistrarMock(), createLogHandlerMock());
+    const server = new LocalSchedulerRuntimeServer(scheduler, createLogHandlerMock());
     await server.stop();
     expect(scheduler.stop).toHaveBeenCalledTimes(1);
   });
